Silence a warning

+++ assembler/src/projects/spades/gap_closer.cpp
@@ -102,7 +102,7 @@ private:
 
 public:
     GapCloserPairedIndexFiller(const GraphPack &gp, omnigraph::de::PairedInfoIndexT<Graph> &paired_index,
-        int max_dist_to_tip, int cnt_libs_to_process)
+                               int max_dist_to_tip, size_t cnt_libs_to_process)
             :  graph_(gp.get<Graph>()), paired_index_(paired_index), buffer_pi_(graph_),
             index_(graph_, gp.workdir()), max_dist_to_tip_(max_dist_to_tip), cnt_libs_to_process_(cnt_libs_to_process) {
         PrepareTipMap(out_tip_map_);
